Proj1/requirements.txt
mysql.connector
glob
os

